Do the first naive take on merging ModalWindow into Dialog
